backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
alembic
passlib[bcrypt]
python-jose
bcrypt==4.0.1
langchain
langchain-openai
langchain-community
langchain-core
python-multipart